[Refactor][quantization]Remove deprecated W4A8Linear/W8A8PDMixMoE schemes and sync W4A8 MoE refactor (#13713)
### What this PR does / why we need it?

- Part of https://github.com/vllm-project/vllm-ascend/issues/13318

Some quantization modes are sunset in this PR, including

- W4A8 Linear
- W4A8 Per-Group MoE(2 weight formats)
- W8A8 PDMix MoE

The reasons are as follows:
- It was a temporary solution and will not have any new evolution plan
in the future.
-    The performance is poor, and there are better alternatives.
- The application scope is small, and the maintenance difficulty is
increased.

### Does this PR introduce _any_ user-facing change?

no

### How was this patch tested?


- vLLM version: v0.26.0
- vLLM main:
https://github.com/vllm-project/vllm/commit/0351e9aa1fdf1a51329d1906881528dfe61fc88e

---------
