+++ .github/workflows/misc/model_dataset_list.json
@@ -224,11 +224,11 @@
       "vllm-ascend/Qwen3-30B-A3B",
       "vllm-ascend/Qwen3-a3B_eagle3",
       "vllm-ascend/Qwen3-30B-A3B-Puring",
+      "vllm-ascend/Qwen3-30B-A3B-quantized.w4a8",
       "vllm-ascend/Qwen3-30B-A3B-W8A8",
       "vllm-ascend/Qwen3-30B-A3B-W8A8-Pruning",
       "vllm-ascend/Qwen3-30B-A3B-W8A8-QuaRot",
       "vllm-ascend/Qwen3-30B-A3B-Instruct-2507-quantized.w8a8",
-      "vllm-ascend/Qwen3-30B-A3B-Instruct-2507-quantized.w4a8",
       "vllm-ascend/Qwen3-32B-W4A4",
       "vllm-ascend/Qwen3-32B-W8A8",
       "vllm-ascend/Qwen3-32B-W8A8-QuaRot",
