+++ docs/source/locale/zh_CN/LC_MESSAGES/developer_guide/Design_Documents/quantization.po
@@ -208,14 +208,14 @@ msgstr ""
 "|"
 
 msgid ""
-"| `W4A8_DYNAMIC`           | INT4   | INT8       | Per-Group          | Per-"
-"Token              | Dynamic | Supports both direct per-channel quantization"
-" to 4-bit and two-step quantization (per-channel to 8-bit then per-group to "
-"4-bit)                                    |"
+"| `W4A8_DYNAMIC`           | INT4   | INT8       | Per-Channel        | Per-"
+"Token              | Dynamic | 4-bit per-channel weight quantization with 8-bit "
+"dynamic per-token activation, supporting msModelSlim and LLM-Compressor "
+"weight formats for MoE layers           |"
 msgstr ""
-"| `W4A8_DYNAMIC`           | INT4   | INT8       | 按组               | 按Token"
-"                | 动态    | 支持直接按通道量化至4位，以及两步量化（先按通道至8位，再按组至4位）"
-"                                    |"
+"| `W4A8_DYNAMIC`           | INT4   | INT8       | 按通道             | 按Token"
+"                | 动态    | 4位按通道权重量化与8位动态按Token激活量化，支持MoE层的"
+"msModelSlim和LLM-Compressor权重格式           |"
 
 msgid ""
 "| `W4A4_FLATQUANT_DYNAMIC` | INT4   | INT4       | Per-Channel        | Per-"
