+++ docs/source/locale/zh_CN/LC_MESSAGES/tutorials/models/Qwen3-Dense.po
@@ -352,9 +352,9 @@ msgid ""
 msgstr "| 场景 | 配置 | NPU数量 | TP | DP | FUSED_MC2 | EP开关 | 异步调度 |"
 
 msgid ""
-"# Qwen3-Dense (Qwen3-0.6B/1.7B/4B/8B/14B/32B, W8A8, W4A8, W4A4, W8A8SC-310)"
+"# Qwen3-Dense (Qwen3-0.6B/1.7B/4B/8B/14B/32B, W8A8, W4A4, W8A8SC-310)"
 msgstr ""
-"# Qwen3-Dense（Qwen3-0.6B/1.7B/4B/8B/14B/32B，W8A8，W4A8，W4A4，W8A8SC-310）"
+"# Qwen3-Dense（Qwen3-0.6B/1.7B/4B/8B/14B/32B，W8A8，W4A4，W8A8SC-310）"
 
 msgid ""
 "Qwen3 is the latest generation of large language models in Qwen series, "
@@ -362,11 +362,11 @@ msgid ""
 " Built upon extensive training, Qwen3 delivers groundbreaking advancements "
 "in reasoning, instruction-following, agent capabilities, and multilingual "
 "support. The Dense variants covered in this document include Qwen3-0.6B, "
-"1.7B, 4B, 8B, 14B, and 32B, along with their quantized versions (W8A8, W4A8,"
-" W4A4, and W8A8SC-310) optimized for Ascend NPU deployment."
+"1.7B, 4B, 8B, 14B, and 32B, along with their quantized versions (W8A8, "
+"W4A4, and W8A8SC-310) optimized for Ascend NPU deployment."
 msgstr ""
 "Qwen3是Qwen系列最新一代的大语言模型，提供全面的密集型和混合专家（MoE）模型套件。基于大规模训练，Qwen3在推理、指令遵循、智能体能力和多语言支持方面取得了突破性进展。本文档涵盖的Dense变体包括Qwen3-0.6B、1.7B、4B、8B、14B和32B，以及针对Ascend"
-" NPU部署优化的量化版本（W8A8、W4A8、W4A4和W8A8SC-310）。"
+" NPU部署优化的量化版本（W8A8、W4A4和W8A8SC-310）。"
 
 msgid "**Quantized Versions for Atlas A2/A3 inference products:**"
 msgstr "**适用于Atlas A2/A3推理产品的量化版本：**"
@@ -386,9 +386,6 @@ msgstr "=== \"Atlas A2推理产品\""
 
 msgid "=== \"Atlas A2 inference products / Atlas A3 inference products\""
 msgstr "=== \"Atlas A2推理产品 / Atlas A3推理产品\""
-
-msgid "Qwen3-8B-W4A8:"
-msgstr "Qwen3-8B-W4A8："
 
 msgid "Qwen3-8B-W8A8SC:"
 msgstr "Qwen3-8B-W8A8SC："
@@ -416,15 +413,15 @@ msgstr "| 低延迟 | 单节点（TP8） | 8（Atlas A3推理产品） | W8A8 | 
 
 msgid ""
 "The Qwen3 Dense models are first supported in v0.8.4rc2. W8A8 quantization "
-"was first supported in v0.8.4rc2, W4A8 quantization is supported since "
-"v0.9.1rc2, and W4A4 is supported since v0.11.0rc1. Atlas 300I DUO uses the "
+"was first supported in v0.8.4rc2, and W4A4 is supported since "
+"v0.11.0rc1. Atlas 300I DUO uses the "
 "W8A8SC-310 quantized weights listed in this tutorial. This document is "
 "validated and written based on **vLLM-Ascend v0.21.0**. All **v0.21.0 and "
 "later versions** can run stably. To use the latest features, it is "
 "recommended to use the latest release candidate or official version."
 msgstr ""
 "Qwen3 "
-"Dense模型首次在v0.8.4rc2中支持。W8A8量化首次在v0.8.4rc2中支持，W4A8量化自v0.9.1rc2起支持，W4A4自v0.11.0rc1起支持。Atlas"
+"Dense模型首次在v0.8.4rc2中支持。W8A8量化首次在v0.8.4rc2中支持，W4A4自v0.11.0rc1起支持。Atlas"
 " 300I DUO使用本教程中列出的W8A8SC-310量化权重。本文档基于**vLLM-Ascend "
 "v0.21.0**验证和编写。所有**v0.21.0及更高版本**均可稳定运行。要使用最新功能，建议使用最新的候选版本或正式版本。"
 
@@ -475,14 +472,6 @@ msgid ""
 msgstr ""
 "| Qwen3-32B | 1个Atlas A3推理产品（64GB × 16），1个Atlas A2推理产品（64GB × 8） | "
 "[下载](https://modelers.cn/models/Modelers_Park/Qwen3-32B) |"
-
-msgid ""
-"| Qwen3-8B-W4A8 | W4A8 | 1 Atlas A3 inference products (64GB × 16) or 1 "
-"Atlas A2 inference products (64GB × 8) | "
-"[Download](https://www.modelscope.cn/models/vllm-ascend/Qwen3-8B-W4A8) |"
-msgstr ""
-"| Qwen3-8B-W4A8 | W4A8 | 1个Atlas A3推理产品（64GB × 16）或1个Atlas A2推理产品（64GB × 8） "
-"| [下载](https://www.modelscope.cn/models/vllm-ascend/Qwen3-8B-W4A8) |"
 
 msgid ""
 "| Qwen3-32B-W4A4 | W4A4 | 1 Atlas A3 inference products (64GB × 16) or 1 "
