+++ tests/e2e/nightly/single_node/models/configs/Qwen3-30B-A3B-W4A8-llm-compressor.yaml
@@ -4,7 +4,7 @@
 
 test_cases:
   - name: "Qwen3-30B-A3B-W4A8-llm-compressor"
-    model: "vllm-ascend/Qwen3-30B-A3B-Instruct-2507-quantized.w4a8"
+    model: "vllm-ascend/Qwen3-30B-A3B-quantized.w4a8"
     envs:
       OMP_PROC_BIND: "false"
       OMP_NUM_THREADS: "1"
